requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3.56
pytesseract>=0.3.8
Pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0
keyboard>=0.13.5
matplotlib>=3.4.3
pandas>=1.3.0
scikit-learn>=0.24.2
yt-dlp>=2023.3.4
psutil>=7.0.0